Refactor join/init values
