+++ pkg/cmd/init/options.go
@@ -3,17 +3,16 @@ package init
 
 import (
 	"k8s.io/cli-runtime/pkg/genericclioptions"
-	operatorv1 "open-cluster-management.io/api/operator/v1"
+	"open-cluster-management.io/clusteradm/pkg/cmd/init/scenario"
 	genericclioptionsclusteradm "open-cluster-management.io/clusteradm/pkg/genericclioptions"
 	"open-cluster-management.io/clusteradm/pkg/helpers/helm"
-	"open-cluster-management.io/clusteradm/pkg/helpers/resourcerequirement"
 )
 
 // Options is holding all the command-line options
 type Options struct {
 	//ClusteradmFlags: The generic options from the clusteradm cli-runtime.
 	ClusteradmFlags *genericclioptionsclusteradm.ClusteradmFlags
-	values          Values
+	values          scenario.Values
 	//The file to output the resources will be sent to the file.
 	outputFile string
 	//If true the bootstrap token will be used instead of the service account token
@@ -46,53 +45,6 @@ type Options struct {
 	output string
 
 	Streams genericclioptions.IOStreams
-}
-
-type BundleVersion struct {
-	// registration image version
-	RegistrationImageVersion string
-	// placement image version
-	PlacementImageVersion string
-	// work image version
-	WorkImageVersion string
-	// operator image version
-	OperatorImageVersion string
-	// addon manager image version
-	AddonManagerImageVersion string
-}
-
-// Values: The values used in the template
-type Values struct {
-	//The values related to the hub
-	Hub Hub `json:"hub"`
-	//bundle version
-	BundleVersion BundleVersion
-
-	// if enable auto approve
-	AutoApprove bool
-
-	// Features is the slice of feature for registration
-	RegistrationFeatures []operatorv1.FeatureGate
-
-	// Features is the slice of feature for work
-	WorkFeatures []operatorv1.FeatureGate
-
-	// Features is the slice of feature for addon manager
-	AddonFeatures []operatorv1.FeatureGate
-
-	// ResourceRequirement is the resource requirement setting for the containers managed by the cluster manager
-	// and the cluster manager operator
-	ResourceRequirement resourcerequirement.ResourceRequirement
-}
-
-// Hub: The hub values for the template
-type Hub struct {
-	//TokenID: A token id allowing the cluster to connect back to the hub
-	TokenID string `json:"tokenID"`
-	//TokenSecret: A token secret allowing the cluster to connect back to the hub
-	TokenSecret string `json:"tokenSecret"`
-	// Registry is the name of the image registry to pull.
-	Registry string `json:"registry"`
 }
 
 func newOptions(clusteradmFlags *genericclioptionsclusteradm.ClusteradmFlags, streams genericclioptions.IOStreams) *Options {
